+++ tool/templates/bloc_source.dart
@@ -1,4 +1,4 @@
-// Version: 3.0.0
+// Version: 4.0.0
 
 import 'dart:async';
 
@@ -9,16 +9,6 @@ import 'package:rivertion/src/internals.dart';
 
 extension SourceStateStreamableExtension<T> on StateStreamable<T> {
   SourceListenable<T> get source => _StateStreamableSource(this);
-
-  @Deprecated('In favour of source.select')
-  SourceListenable<R> select<R>(R Function(T state) selector) => source.select(selector);
-
-  @Deprecated('In favour of source.selectWith')
-  SourceListenable<R> selectWith<R, A>(A arg, R Function(A arg, T state) selector) =>
-      source.selectWith(arg, selector);
-
-  @Deprecated('In favour of source.where')
-  SourceListenable<T> where(bool Function(T previous, T next) condition) => source.where(condition);
 }
 
 final class _StateStreamableSource<T> extends SourceListenable<T> {
@@ -30,7 +20,6 @@ final class _StateStreamableSource<T> extends SourceListenable<T> {
   SourceSubscription<T> listen(SourceListener<T> onChange) {
     var current = _stateStreamable.state;
 
-    // ignore: cancel_subscriptions
     final streamSubscription = _stateStreamable.stream.listen((next) {
       final previous = current;
       current = next;
@@ -45,7 +34,7 @@ final class _StateStreamableSource<T> extends SourceListenable<T> {
       identical(this, other) ||
       other is _StateStreamableSource<T> &&
           runtimeType == other.runtimeType &&
-          _stateStreamable == other._stateStreamable;
+          identical(_stateStreamable, other._stateStreamable);
 
   @override
   int get hashCode => _stateStreamable.hashCode;