+++ tool/templates/riverpod_ref_extensions.dart
@@ -1,0 +1,32 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+// ignore: implementation_imports
+import 'package:rivertion/src/internals.dart';
+
+extension SourcesRefExtension on Ref {
+  T watchSource<T>(Source<T> source) {
+    final subscription = source.listenable.listen((_, _) => invalidateSelf(asReload: true));
+    onDispose(subscription.cancel);
+    return subscription.read();
+  }
+
+  SourceSubscription<T> listenSource<T>(Source<T> source, SourceListener<T> listener) {
+    final subscription = source.listenable.listen(listener);
+    final dispositionRemover = onDispose(subscription.cancel);
+    return SourceSubscriptionBuilder(subscription.read, () {
+      dispositionRemover();
+      subscription.cancel();
+    });
+  }
+
+  SourceSubscription<T> listenSourceImmediately<T>(
+    Source<T> source,
+    SourceImmediatelyListener<T> listener,
+  ) {
+    final subscription = source.listenable.listenImmediately(listener);
+    final dispositionRemover = onDispose(subscription.cancel);
+    return SourceSubscriptionBuilder(subscription.read, () {
+      dispositionRemover();
+      subscription.cancel();
+    });
+  }
+}